Add configuration folders as deb conffiles
`CONFIG | NOREPLACE` doesn't seem to be behaving correctly, if we use confgurationFile() then debian upgrades are aware of them.

I've tested this from gradle build to gradle build AND from cloudbees to gradle build and it looks like conf file pathing is the same in each. Just in case I've missed something, please test as well.