improve by pgrep the fetching of the openhab PID
The present design of the findpid function involves a 'grep' on the output of the 'ps aux' command. This approach can work but it is depending on the settings of the terminal invoking the ps command (in particular the width of the terminal set in the COLUMNS environment variable). A far more elegant approach is proposed here (actually wisely suggested by BClark09, and then successfully tested) that makes use of 'pgrep' that in turns allows to significantly simplify the function and make it 100% reliable regardless of the calling environment.

Signed-off-by: Olivier Colin oliviercommelarbre@gmail.com